Fix undefined add_show_tools_partial.
In upgrading a BL-6.14.1 site to Bl-7 I got the an undefined method
add_show_tools_partial for this gem.

This function will be deprecated in the next release. I'm not sure why
I got the errors, but updating the code to use the non deprecated
variant, fixed the issue for me.

This commit fixes this issue.